Cleanup: SingleCLip and LinearBlend jobs
