+++ Runtime/Components/LinearBlendStateMachineState.cs
@@ -6,11 +6,11 @@ namespace DMotion
     [BurstCompile]
     internal struct LinearBlendStateMachineState : IBufferElementData
     {
+        internal byte PlayableId;
         internal BlobAssetReference<StateMachineBlob> StateMachineBlob;
         internal short StateIndex;
-        internal byte PlayableId;
         internal readonly ref AnimationStateBlob StateBlob => ref StateMachineBlob.Value.States[StateIndex];
-        internal readonly ref LinearBlendStateBlob AsLinearBlend =>
+        internal readonly ref LinearBlendStateBlob LinearBlendBlob =>
             ref StateMachineBlob.Value.LinearBlendStates[StateBlob.StateIndex];
     }
 }
