+++ Runtime/Systems/SingleClipStateJobs.cs
@@ -1,41 +1,15 @@
 ﻿using System.Runtime.CompilerServices;
 using Unity.Burst;
-using Unity.Collections;
-using Unity.Collections.LowLevel.Unsafe;
 using Unity.Entities;
 
 namespace DMotion
 {
     [BurstCompile]
-    internal partial struct UpdateSingleClipStatesJob : IJobEntityBatch
+    internal partial struct UpdateSingleClipStatesJob
     {
         internal float DeltaTime;
-        [NativeDisableContainerSafetyRestriction]
-        internal BufferTypeHandle<ClipSampler> ClipSamplersHandle;
-        
-        [ReadOnly]
-        internal BufferTypeHandle<SingleClipState> SingleClipStatesHandle;
-        [ReadOnly]
-        internal BufferTypeHandle<PlayableState> PlayableStatesHandle;
-
-        public void Execute(ArchetypeChunk batchInChunk, int batchIndex)
-        {
-            var clipSamplersAccessor = batchInChunk.GetBufferAccessor(ClipSamplersHandle);
-            var singleStatesAccessor = batchInChunk.GetBufferAccessor(SingleClipStatesHandle);
-            var playableStatesAccessor = batchInChunk.GetBufferAccessor(PlayableStatesHandle);
-            
-            for(var i = 0; i < batchInChunk.Count; i++)
-            {
-                var clipSamplers = clipSamplersAccessor[i];
-                var singleStates = singleStatesAccessor[i];
-                var playables = playableStatesAccessor[i];
-                
-                Execute(DeltaTime, ref clipSamplers, singleStates, playables);
-            }
-        }
-        
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static void Execute(float dt,
+        internal void Execute(
             ref DynamicBuffer<ClipSampler> clipSamplers,
             in DynamicBuffer<SingleClipState> singleClipStates,
             in DynamicBuffer<PlayableState> playableStates
@@ -46,35 +20,17 @@ namespace DMotion
                 if (playableStates.TryGetWithId(singleClipStates[i].PlayableId, out var playable))
                 {
                     SingleClipStateUtils
-                        .UpdateSamplers(singleClipStates[i], dt, playable, ref clipSamplers);
+                        .UpdateSamplers(singleClipStates[i], DeltaTime, playable, ref clipSamplers);
                 }
             }
         }
     }
     
     [BurstCompile]
-    internal partial struct CleanSingleClipStatesJob : IJobEntityBatch
+    internal partial struct CleanSingleClipStatesJob
     {
-        internal BufferTypeHandle<SingleClipState> SingleClipStatesHandle;
-        [ReadOnly]
-        internal BufferTypeHandle<PlayableState> PlayableStatesHandle;
-
-        public void Execute(ArchetypeChunk batchInChunk, int batchIndex)
-        {
-            var singleStatesAccessor = batchInChunk.GetBufferAccessor(SingleClipStatesHandle);
-            var playableStatesAccessor = batchInChunk.GetBufferAccessor(PlayableStatesHandle);
-            
-            for(var i = 0; i < batchInChunk.Count; i++)
-            {
-                var singleStates = singleStatesAccessor[i];
-                var playables = playableStatesAccessor[i];
-                
-                Execute(ref singleStates, playables);
-            }
-        }
-        
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static void Execute(
+        internal void Execute(
             ref DynamicBuffer<SingleClipState> singleClipStates,
             in DynamicBuffer<PlayableState> playableStates
         )
